+++ Zilon.Core/CDT.LAST.MonoGameClient/Screens/GlobeSelectionScreen.cs
@@ -17,7 +17,7 @@ namespace CDT.LAST.MonoGameClient.Screens
 {
     internal class GlobeSelectionScreen : GameSceneBase
     {
-        private readonly Button _generateButton;
+        private readonly TextButton _generateButton;
         private readonly SpriteBatch _spriteBatch;
         private bool _generationWasStarted;
 
@@ -28,7 +28,7 @@ namespace CDT.LAST.MonoGameClient.Screens
             var buttonTexture = game.Content.Load<Texture2D>("Sprites/ui/button");
             var font = Game.Content.Load<SpriteFont>("Fonts/Main");
 
-            _generateButton = new Button(UiResources.GenerateGlobeButtonTitle, buttonTexture, font,
+            _generateButton = new TextButton(UiResources.GenerateGlobeButtonTitle, buttonTexture, font,
                 new Rectangle(150, 150, 100, 20));
 
             _generateButton.OnClick += GenerateButtonClickHandlerAsync;
